pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-scraping-template"
version = "1.0.0"
description = "A reusable, configurable web scraping template with Playwright."
authors = [{name = "thoni-dev"}]
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "playwright>=1.42.0",
    "beautifulsoup4>=4.12.3",
    "pandas>=2.2.0",
    "pyyaml>=6.0.1",
    "requests>=2.32.0",
    "openpyxl>=3.1.2",
] 